feat: removed notempty enforcement in post request for date_published in new_article.php

+++ delete_article.php
@@ -10,7 +10,7 @@ $conn = connectDB();
 // READING or RETRIEVING from the database to get specific article post by their ids
 if (isset($_GET['id'])){
 
-    // checks if the article's id exits in the database, then returns an associative array, which stores in $article variable
+    // checks if the article's id exits in the database, then returns an associative array
     $article = getArticle($conn, $_GET['id']); 
 
     if ($article){
@@ -31,27 +31,32 @@ delete. The delete button is the only condition that can make that article to be
 */
 
 if ($_SERVER["REQUEST_METHOD"] == "POST"){
-    // Delete the data into the database server
-    $sql = "DELETE FROM article 
-            WHERE id = ?";
+    if (isset($_POST["delete"])){
+        // Delete the data in the database server by its id row
+        $sql = "DELETE FROM article 
+                WHERE id = ?";
 
-    $stmt = mysqli_prepare($conn, $sql);
+        // Prepares an SQL statement for execution
+        $stmt = mysqli_prepare($conn, $sql);
 
-    if ($stmt === false){
-        echo mysqli_error($conn);
-    } else {
-        // i - integer, d - double, s - string
-        mysqli_stmt_bind_param($stmt, "i", $id);
-
-        $results = mysqli_stmt_execute($stmt);
-
-        // checking for errors, if none, then redirect the user to the new article page
-        if ($results === false){
-            echo mysqli_stmt_error($stmt);
+        if ($stmt === false){
+            echo mysqli_error($conn);
         } else {
-            // it is more advisable to use absolute paths below than relative path
-            header("Location: http://localhost/lexispress_cms-app/index.php"); 
-            exit;
+            // i - integer, d - double, s - string
+            // Bind variables for the parameter markers in the SQL statement prepared
+            mysqli_stmt_bind_param($stmt, "i", $id);
+
+            // Executes a prepared statement
+            $results = mysqli_stmt_execute($stmt);
+
+            // checking for errors, if none, then redirect the user to the new article page
+            if ($results === false){
+                echo mysqli_stmt_error($stmt);
+            } else {
+                // it is more advisable to use absolute paths below than relative path
+                header("Location: http://localhost/lexispress_cms-app/index.php"); 
+                exit;
+            }
         }
     }
 }
